Fixed 2 bugs (v1.2.1)
- Fixed a bug where stringifying non-printable unicode characters would
throw a ClassCastException (in JSONStringify).
- Fixed a bug where checking for invalid unicode surrogate pairs would
not work as intended (in JSONParser).

+++ src/main/java/at/syntaxerror/json5/JSONStringify.java
@@ -328,7 +328,7 @@ public class JSONStringify {
 				case Character.SURROGATE:
 				case Character.UNASSIGNED:
 					quoted.append("\\u");
-					quoted.append(String.format("%04X", c));
+					quoted.append(String.format("%04X", (int) c));
 					break;
 				default:
 					quoted.append(c);
